@@ -1,10 +1,12 @@
-require File.expand_path('../boot', __FILE__)
+require_relative 'boot'
 
 require 'rails/all'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
-Bundler.require(:default, Rails.env)
+Bundler.require(*Rails.groups)
+
+ActiveSupport.halt_callback_chains_on_return_false = false
 
 module SnowSchoolers
   class Application < Rails::Application
