@@ -1,4 +1,4 @@
-SnowSchoolers::Application.configure do
+Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
   # In the development environment your application's code is reloaded on
@@ -18,11 +18,24 @@ SnowSchoolers::Application.configure do
 
   # Show full error reports and disable caching.
   config.consider_all_requests_local       = true
-  config.action_controller.perform_caching = false
+
+
+  if Rails.root.join('tmp/caching-dev.txt').exist?
+     config.action_controller.perform_caching = true
+     config.cache_store = :memory_store
+     config.public_file_server.headers = {
+       'Cache-Control' => 'public, max-age=172800'
+     }
+  else
+     config.action_controller.perform_caching = false
+ 
+     config.cache_store = :null_store
+  end
 
   # Don't care if the mailer can't send.
   config.action_mailer.raise_delivery_errors = false
 
+  config.action_mailer.perform_caching = false
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
 
@@ -34,12 +47,13 @@ SnowSchoolers::Application.configure do
   # number of complex assets.
   config.assets.debug = true
 
-
+  #Suppress logger output for asset requests
+  config.assets.quiet = true
   #GMAIL CONFIG
   config.action_mailer.default_url_options = { :host => 'snowschoolers.com' }
   config.action_mailer.delivery_method = :smtp
   config.action_mailer.perform_deliveries = true
-  # config.action_mailer.perform_deliveries = false
+   #config.action_mailer.perform_deliveries = false
   config.action_mailer.raise_delivery_errors = true
   config.action_mailer.default :charset => "utf-8"
   config.action_mailer.smtp_settings = {
@@ -59,4 +73,5 @@ SnowSchoolers::Application.configure do
   :bucket => 'citizen-debate'
   }
 
+  # config.file_watcher = ActiveSupport::EventedFileUpdateChecker
 end
