@@ -1,4 +1,4 @@
-SnowSchoolers::Application.configure do
+Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
   # Code is not reloaded between requests.
@@ -19,9 +19,10 @@ SnowSchoolers::Application.configure do
   # For large-scale production use, consider using a caching reverse proxy like nginx, varnish or squid.
   # config.action_dispatch.rack_cache = true
 
-  # Disable Rails's static asset server (Apache or nginx will already do this).
-  config.serve_static_assets = false
-
+  
+  # Disable serving static files from the `/public` folder by default since
+	  # Apache or NGINX already handles this.
+  config.public_file_server.enabled = ENV['RAILS_SERVE_STATIC_FILES'].present?
   # Compress JavaScripts and CSS.
   config.assets.js_compressor = :uglifier
   # config.assets.css_compressor = :sass
@@ -38,7 +39,10 @@ SnowSchoolers::Application.configure do
   # Specifies the header that your server uses for sending files.
   # config.action_dispatch.x_sendfile_header = "X-Sendfile" # for apache
   # config.action_dispatch.x_sendfile_header = 'X-Accel-Redirect' # for nginx
-
+  # Mount Action Cable outside main process or domain
+	# config.action_cable.mount_path = nil
+	# config.action_cable.url = 'wss://example.com/cable'
+	# config.action_cable.allowed_request_origins = [ 'http://example.com', /http:\/\/example.*/ ]
   # Force all access to the app over SSL, use Strict-Transport-Security, and use secure cookies.
   # config.force_ssl = true
 
@@ -103,4 +107,11 @@ SnowSchoolers::Application.configure do
     password: ENV["GMAIL_PASSWORD"]
   }
 
+  if ENV["RAILS_LOG_TO_STDOUT"].present?
+    logger           = ActiveSupport::Logger.new(STDOUT)
+    logger.formatter = config.log_formatter
+    config.logger = ActiveSupport::TaggedLogging.new(logger)
+  end
+  # Do not dump schema after migrations.
+  config.active_record.dump_schema_after_migration = false
 end
