@@ -1,3 +1,3 @@
 # Be sure to restart your server when you modify this file.
 
-SnowSchoolers::Application.config.session_store :cookie_store, key: '_snow-schoolers_session'
+Rails.application.config.session_store :cookie_store, key: '_snow_schoolers_session'
